fix(ainic): add copyright headers and fix data type mismatches
AMD SMI returns uint64_t for NIC statistics but the AI NIC sampling
code was using uint32_t, which could cause data truncation for
high-throughput network counters.

Changes:
- Add SPDX copyright headers to AI NIC source files
- Change NIC stats fields from uint32_t to uint64_t to match AMD SMI API
- Remove redundant amdsmi.h include (already included via header)

+++ projects/rocprofiler-systems/source/lib/rocprof-sys/library/amd_smi_ainic.cpp
@@ -1,10 +1,12 @@
+// Copyright (c) Advanced Micro Devices, Inc.
+// SPDX-License-Identifier: MIT
+
 #include "library/amd_smi_ainic.hpp"
 
 #include "core/agent_manager.hpp"
 #include "core/trace_cache/cache_manager.hpp"
 #include "core/trace_cache/cacheable.hpp"
 #include "core/trace_cache/sample_type.hpp"
-#include <amd_smi/amdsmi.h>
 #include <cstdint>
 #if defined(NDEBUG)
 #    undef NDEBUG
@@ -229,12 +231,12 @@ nic_data::post_process(size_t nic_index)
         uint64_t _ts = itr.m_ts;
         if(!_thread_info->is_valid_time(_ts)) continue;
 
-        uint32_t _rx_rdma_cnp_pkts = itr._rx_rdma_cnp_pkts;
-        uint32_t _tx_rdma_cnp_pkts = itr._tx_rdma_cnp_pkts;
-        uint32_t _rx_ucast_bytes   = itr._rx_ucast_bytes;
-        uint32_t _tx_ucast_bytes   = itr._tx_ucast_bytes;
-        uint32_t _rx_ucast_pkts    = itr._rx_ucast_pkts;
-        uint32_t _tx_ucast_pkts    = itr._tx_ucast_pkts;
+        uint64_t _rx_rdma_cnp_pkts = itr._rx_rdma_cnp_pkts;
+        uint64_t _tx_rdma_cnp_pkts = itr._tx_rdma_cnp_pkts;
+        uint64_t _rx_ucast_bytes   = itr._rx_ucast_bytes;
+        uint64_t _tx_ucast_bytes   = itr._tx_ucast_bytes;
+        uint64_t _rx_ucast_pkts    = itr._rx_ucast_pkts;
+        uint64_t _tx_ucast_pkts    = itr._tx_ucast_pkts;
 
         counter_track::emplace(nic_index, addendum("RX CNP PKTS"), "packets");
         counter_track::emplace(nic_index, addendum("TX CNP PKTS"), "packets");
